feat: add ShareLinkSettingsModal component with types and utility functions
- Introduced ShareLinkSettingsModal with types for share link settings.
- Added utility functions for handling selection settings and expiration logic.
- Refactored ProjectPage and ShareLinkDetailPage to utilize new components and utilities.
- Enhanced ShareLinksDashboardPage with improved structure and utility functions for better readability and maintainability.
