+++ frontend/src/components/project-page/types.ts
@@ -1,0 +1,5 @@
+export type GalleryDraft = {
+  name: string;
+  shooting_date: string;
+  project_visibility: 'listed' | 'direct_only';
+};
